LayoutTests/http/tests/security/xssAuditor/script-tag-post-expected.txt

CONSOLE MESSAGE: Refused to execute a JavaScript script. Source code of script found within request.
